fix: sonar duplicated code
